use SwiftAudioPlayer replace AVAudioPlayer
