+++ example/ios/Podfile
@@ -1,16 +1,16 @@
 # Uncomment this line to define a global platform for your project
-# platform :ios, '9.0'
+platform :ios, '10.0'
 
 # CocoaPods analytics sends network stats synchronously affecting flutter build latency.
 ENV['COCOAPODS_DISABLE_STATS'] = 'true'
 
 project 'Runner', {
-  'Debug' => :debug,
-  'Profile' => :release,
-  'Release' => :release,
+    'Debug' => :debug,
+    'Profile' => :release,
+    'Release' => :release,
 }
 
-def parse_KV_file(file, separator='=')
+def parse_KV_file(file, separator = '=')
   file_abs_path = File.expand_path(file)
   if !File.exists? file_abs_path
     return [];
@@ -19,7 +19,7 @@ def parse_KV_file(file, separator='=')
   skip_line_start_symbols = ["#", "/"]
   File.foreach(file_abs_path) do |line|
     next if skip_line_start_symbols.any? { |symbol| line =~ /^\s*#{symbol}/ }
-    plugin = line.split(pattern=separator)
+    plugin = line.split(pattern = separator)
     if plugin.length == 2
       podname = plugin[0].strip()
       path = plugin[1].strip()
