+++ ios/Classes/PlayMode.swift
@@ -1,0 +1,11 @@
+//
+// Created by yangbin on 2020/2/22.
+//
+
+import Foundation
+
+enum PlayMode: Int {
+    case shuffle = 0
+    case single
+    case sequence
+}
